Refactoring: Move parse_labels to parse module
