+++ src/parse.rs
@@ -10,6 +10,8 @@ mod resource_path;
 pub(crate) mod widely_used_args;
 
 pub(crate) use byte_size::ByteSizeParser;
-pub(crate) use helpers::{fetch_and_filter_entries, parse_query_params, parse_time, QueryParams};
+pub(crate) use helpers::{
+    fetch_and_filter_entries, parse_labels, parse_query_params, parse_time, QueryParams,
+};
 pub(crate) use quota_type::QuotaTypeParser;
 pub(crate) use resource_path::{Resource, ResourcePathParser};
