+++ src/parse/helpers.rs
@@ -8,8 +8,9 @@ use crate::context::CliContext;
 use chrono::{DateTime, Local, LocalResult, NaiveDate, NaiveDateTime, TimeZone, Utc};
 use clap::parser::MatchesError;
 use clap::ArgMatches;
-use reduct_rs::{Bucket, EntryInfo};
+use reduct_rs::{Bucket, EntryInfo, Labels};
 use serde_json::Value;
+use std::collections::HashMap;
 use std::time::Duration;
 
 pub(crate) async fn fetch_and_filter_entries(
@@ -179,6 +180,46 @@ fn timestamp_micros_to_u64(micros: i64, time_str: &str) -> anyhow::Result<u64> {
             time_str
         )
     })
+}
+
+/// Parse labels from either key=value,key=value format or JSON object format
+pub(crate) fn parse_labels(labels_str: &str) -> anyhow::Result<Labels> {
+    let trimmed = labels_str.trim();
+
+    // Try to parse as JSON first
+    if trimmed.starts_with('{') {
+        let labels_map: HashMap<String, String> = serde_json::from_str(trimmed)
+            .map_err(|e| anyhow::anyhow!("Failed to parse labels as JSON: {}", e))?;
+        return Ok(labels_map);
+    }
+
+    // Parse as key=value,key=value format
+    let mut labels = HashMap::new();
+    for pair in trimmed.split(',') {
+        let pair = pair.trim();
+        if pair.is_empty() {
+            continue;
+        }
+
+        let parts: Vec<&str> = pair.splitn(2, '=').collect();
+        if parts.len() != 2 {
+            return Err(anyhow::anyhow!(
+                "Invalid label format '{}': expected key=value",
+                pair
+            ));
+        }
+
+        let key = parts[0].trim();
+        let value = parts[1].trim();
+
+        if key.is_empty() {
+            return Err(anyhow::anyhow!("Label key cannot be empty in '{}'", pair));
+        }
+
+        labels.insert(key.to_string(), value.to_string());
+    }
+
+    Ok(labels)
 }
 
 #[derive(Clone, Debug)]
@@ -656,6 +697,78 @@ mod test {
 
             assert_eq!(query_params.parallel, 3);
             assert_eq!(query_params.ttl, Duration::from_secs(36));
+        }
+    }
+
+    mod parse_labels {
+
+        #[test]
+        fn test_parse_labels_key_value_format() {
+            let labels = parse_labels("key1=value1,key2=value2").unwrap();
+            assert_eq!(labels.len(), 2);
+            assert_eq!(labels.get("key1"), Some(&"value1".to_string()));
+            assert_eq!(labels.get("key2"), Some(&"value2".to_string()));
+        }
+
+        #[test]
+        fn test_parse_labels_key_value_with_spaces() {
+            let labels = parse_labels(" key1 = value1 , key2 = value2 ").unwrap();
+            assert_eq!(labels.len(), 2);
+            assert_eq!(labels.get("key1"), Some(&"value1".to_string()));
+            assert_eq!(labels.get("key2"), Some(&"value2".to_string()));
+        }
+
+        #[test]
+        fn test_parse_labels_json_format() {
+            let labels = parse_labels(r#"{"key1":"value1","key2":"value2"}"#).unwrap();
+            assert_eq!(labels.len(), 2);
+            assert_eq!(labels.get("key1"), Some(&"value1".to_string()));
+            assert_eq!(labels.get("key2"), Some(&"value2".to_string()));
+        }
+
+        #[test]
+        fn test_parse_labels_empty_value() {
+            let labels = parse_labels("key1=,key2=value2").unwrap();
+            assert_eq!(labels.len(), 2);
+            assert_eq!(labels.get("key1"), Some(&"".to_string()));
+            assert_eq!(labels.get("key2"), Some(&"value2".to_string()));
+        }
+
+        #[test]
+        fn test_parse_labels_value_with_equals() {
+            let labels = parse_labels("key1=value=with=equals").unwrap();
+            assert_eq!(labels.len(), 1);
+            assert_eq!(labels.get("key1"), Some(&"value=with=equals".to_string()));
+        }
+
+        #[test]
+        fn test_parse_labels_invalid_format() {
+            let result = parse_labels("invalid");
+            assert!(result.is_err());
+            assert!(result
+                .unwrap_err()
+                .to_string()
+                .contains("expected key=value"));
+        }
+
+        #[test]
+        fn test_parse_labels_empty_key() {
+            let result = parse_labels("=value");
+            assert!(result.is_err());
+            assert!(result
+                .unwrap_err()
+                .to_string()
+                .contains("key cannot be empty"));
+        }
+
+        #[test]
+        fn test_parse_labels_invalid_json() {
+            let result = parse_labels(r#"{"key1":"value1""#);
+            assert!(result.is_err());
+            assert!(result
+                .unwrap_err()
+                .to_string()
+                .contains("Failed to parse labels as JSON"));
         }
     }
 }
